+++ web/apps/photos/src/components/UploadProgress.tsx
